Format two numbers with a joining comma, rather than as a single range
In longer range strings, it would be weird to see two neighbouring
numbers as a single range (eg "1,2,3,5" vs "1-2,3-5x2").
It looks a bit odd in a two-length range, just because seeing "X-Y"
is more common (eg "1,2####" vs "1-2####").
But the implementation of normalisation is quite a bit easier this way,
and having only two digits in the range is such a rare edge case
that the complexity doesn't seem worth it.